@@ -1,5 +1,6 @@
 'use strict';
-import React, { PropTypes as T } from 'react';
+import React from 'react';
+import { PropTypes as T } from 'prop-types';
 
 class FileInput extends React.PureComponent {
   constructor (props) {
