@@ -1,6 +1,7 @@
 'use strict';
 import React from 'react';
-import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
+import { createPortal } from 'react-dom';
+import { CSSTransition } from 'react-transition-group';
 
 // Minimum time the loading is visible.
 const MIN_TIME = 512;
@@ -17,23 +18,18 @@ var theGlobalLoading = null;
 // all were hidden.
 let theGlobalLoadingCount = 0;
 
-// Decrement function to ensure the count never goes below zero.
-let decrementCount = () => {
-  if (--theGlobalLoadingCount < 0) theGlobalLoadingCount = 0;
-  return theGlobalLoadingCount;
-};
+export class GlobalLoading extends React.Component {
+  constructor (props) {
+    super(props);
+    this.componentAddedBodyClass = false;
 
-const GlobalLoading = React.createClass({
-  componentAddedBodyClass: false,
-
-  getInitialState: function () {
-    return {
+    this.state = {
       showTimestamp: 0,
       revealed: false
     };
-  },
+  }
 
-  toggleBodyClass: function (revealed) {
+  toggleBodyClass (revealed) {
     let bd = document.getElementsByTagName('body')[0];
     if (revealed) {
       this.componentAddedBodyClass = true;
@@ -43,60 +39,70 @@ const GlobalLoading = React.createClass({
       this.componentAddedBodyClass = false;
       bd.classList.remove('unscrollable-y');
     }
-  },
+  }
 
-  componentDidMount: function () {
+  componentDidMount () {
     if (theGlobalLoading !== null) {
       throw new Error('<GlobalLoading /> component was already mounted. Only 1 is allowed.');
     }
     theGlobalLoading = this;
     this.toggleBodyClass(this.state.revealed);
-  },
-
-  componentDidUpdate: function () {
-    this.toggleBodyClass(this.state.revealed);
-  },
-
-  componentWillUnmount: function () {
-    this.toggleBodyClass(false);
-  },
-
-  renderLoading: function () {
-    return (
-      <div className='loading-pane'>
-        <div className='spinner'>
-          <div className='spinner__bounce'></div>
-          <div className='spinner__bounce'></div>
-          <div className='spinner__bounce'></div>
-        </div>
-      </div>
-    );
-  },
-
-  render: function () {
-    return (
-      <ReactCSSTransitionGroup
-        component='div'
-        transitionName='loading-pane'
-        transitionEnterTimeout={300}
-        transitionLeaveTimeout={300} >
-
-        {this.state.revealed ? this.renderLoading() : null}
-
-      </ReactCSSTransitionGroup>
-    );
   }
-});
 
-export function showGlobalLoading () {
-  showGlobalLoadingCounted(0);
+  componentDidUpdate () {
+    this.toggleBodyClass(this.state.revealed);
+  }
+
+  componentWillUnmount () {
+    this.toggleBodyClass(false);
+    theGlobalLoading = null;
+  }
+
+  render () {
+    return createPortal((
+      <CSSTransition
+        in={this.state.revealed}
+        unmountOnExit={true}
+        appear={true}
+        classNames='loading-pane'
+        timeout={{ enter: 300, exit: 300 }}>
+
+        <div className='loading-pane'>
+          <div className='spinner'>
+            <div className='spinner__bounce'></div>
+            <div className='spinner__bounce'></div>
+            <div className='spinner__bounce'></div>
+          </div>
+        </div>
+
+      </CSSTransition>
+    ), document.body);
+  }
 }
 
-export function hideGlobalLoading (force = false) {
-  hideGlobalLoadingCounted();
-}
-
-export function showGlobalLoadingCounted (count = null) {
+/**
+ * Show a global loading.
+ * The loading has a minimum visible time defined by the MIN_TIME constant.
+ * This will prevent flickers in the interface when the action is very fast.
+ * @param  {Number} count Define how many loadings to show. This will not
+ *                        show multiple loadings on the page but will increment
+ *                        a counter. This is helpful when there are many actions
+ *                        that require a loading.
+ *                        The global loading will only be dismissed once all
+ *                        counters shown are hidden.
+ *                        Each function call will increment the counter.
+ *                        Default 1
+ * @example
+ * showGlobalLoading()
+ * // Counter set to 1
+ * showGlobalLoading(3)
+ * // Counter set to 4
+ * hideGlobalLoading()
+ * // Counter is now 3
+ * hideGlobalLoading(3)
+ * // Counter is now 0 and the loading is dismissed.
+ */
+export function showGlobalLoading (count = 1) {
   if (theGlobalLoading === null) {
     throw new Error('<GlobalLoading /> component not mounted');
   }
@@ -104,7 +110,7 @@ export function showGlobalLoadingCounted (count = null) {
     clearTimeout(hideTimeout);
   }
 
-  theGlobalLoadingCount = count === null ? theGlobalLoadingCount + 1 : count;
+  theGlobalLoadingCount += count;
 
   theGlobalLoading.setState({
     showTimestamp: Date.now(),
@@ -112,31 +118,54 @@ export function showGlobalLoadingCounted (count = null) {
   });
 }
 
-export function hideGlobalLoadingCounted (force = false) {
+/**
+ * Hides a global loading
+ * @param  {Number} count Define how many loadings to hide. Using 0 or any
+ *                        negative number will immediately dismiss the loading
+ *                        without waiting for the minimum display time.
+ *                        Default 1
+ * @param {function} cb   Callback called after the loading is hidden.
+ *
+ * @example
+ * showGlobalLoading()
+ * // Counter set to 1
+ * showGlobalLoading(3)
+ * // Counter set to 4
+ * hideGlobalLoading()
+ * // Counter is now 3
+ * hideGlobalLoading(3)
+ * // Counter is now 0 and the loading is dismissed.
+ */
+export function hideGlobalLoading (count = 1, cb = () => {}) {
   if (theGlobalLoading === null) {
     throw new Error('<GlobalLoading /> component not mounted');
   }
 
-  const hide = () => theGlobalLoading.setState({revealed: false});
+  if (typeof count === 'function') {
+    cb = count;
+    count = 1;
+  }
 
-  if (force) {
+  const hide = () => { theGlobalLoading.setState({revealed: false}); cb(); };
+
+  // Using 0 or negative numbers results in the loading being
+  // immediately dismissed.
+  if (count < 1) {
     theGlobalLoadingCount = 0;
     return hide();
   }
 
+  // Decrement counter by given amount without going below 0.
+  theGlobalLoadingCount -= count;
+  if (theGlobalLoadingCount < 0) theGlobalLoadingCount = 0;
+
   let time = theGlobalLoading.state.showTimestamp;
   let diff = Date.now() - time;
   if (diff >= MIN_TIME) {
-    if (!decrementCount()) return hide();
+    if (theGlobalLoadingCount === 0) return hide();
   } else {
     hideTimeout = setTimeout(() => {
-      if (!decrementCount()) return hide();
+      if (theGlobalLoadingCount === 0) return hide();
     }, MIN_TIME - diff);
   }
 }
-
-module.exports = {
-  GlobalLoading,
-  showGlobalLoading,
-  hideGlobalLoading
-};
