@@ -1,1 +1,2 @@
-module.exports = require('react-scrollbar/dist/no-css').default;
+console.log('value', require('react-scrollbar').default);
+module.exports = require('react-scrollbar').default;
